+++ tests/color_test.py
@@ -42,8 +42,8 @@ def test_eq_withother(color):
     assert color != Random
 
 
-def test_rgba32(color):
-    assert color.rgba32() == 0x007FFFFF
+def test_argb32(color):
+    assert color.argb32() == 0xFF007FFF
 
 
 def test_darker(color):
@@ -126,12 +126,12 @@ def test_random_defaults(monkeypatch):
         assert c.to_tuple()[:-1] in list(Color._color_defaults.values()) + list(Color._grayscale_defaults.values())
 
 
-def test_from_rgba32(color):
-    assert Color.from_rgba32(0x007FFFFF) == color
+def test_from_argb32(color):
+    assert Color.from_argb32(0xFF007FFF) == color
 
 
-def test_from_rgba32_0():
-    assert Color.from_rgba32(0) == Color(0, 0, 0, 0)
+def test_from_argb32_0():
+    assert Color.from_argb32(0) == Color(0, 0, 0, 0)
 
 
 def test_from_hex(color):